% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivTest_FixedMargin.R
\name{createEquivTestSmpl}
\alias{createEquivTestSmpl}
\title{Create summary information of a dataset}
\usage{
createEquivTestSmpl(smpl)
}
\arguments{
\item{smpl}{a vector representing the dataset}
}
\value{
a list of objects summarizing the dataset
}
\description{
Create a list of summary statistics of a dataset for equivalence test.
}
\examples{
vecT = rnorm(n=20)
s = createEquivTestSmpl(vecT)

}
