% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivTest_FixedMargin.R
\name{dfSatterthwaite}
\alias{dfSatterthwaite}
\title{Compute the Satterthwaite approximation of degree of freedom for t distribution}
\usage{
dfSatterthwaite(s1, n1, n1s, s2, n2, n2s)
}
\arguments{
\item{s1}{sample standard deviation for group 1}

\item{n1}{sample size for group 1}

\item{n1s}{adjusted sample size for group 1}

\item{s2}{sample standard deviation  for group 2}

\item{n2}{sample size for group 2}

\item{n2s}{adjusted sample size for group 2}
}
\value{
degree of freedom
}
\description{
Compute the Satterthwaite approximation of degree of freedom for t distribution.
}
