% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivTest_MWCMLE.R
\name{equivTestMWCMLE}
\alias{equivTestMWCMLE}
\title{Equivalence test by Modified Wald test with standard error estimated by  RMLE (MWCMLE)}
\usage{
equivTestMWCMLE(vecT, vecR, alpha = 0.05, marginX = 1.7,
  method = "MWCMLE")
}
\arguments{
\item{vecT}{the sample data for test product, can be a vector of observed values or a list returned by \code{createEquivTestSmpl}}

\item{vecR}{the sample data for reference product, can be a vector of observed values or a list returned by \code{createEquivTestSmpl}}

\item{alpha}{the nominal size, default = 0.05}

\item{marginX}{the margin multiplier, default = 1.5}

\item{method}{a string indicating the method used in the equivalence test.}
}
\value{
a list containing the test result
}
\description{
Equivalence test by Modified Wald test with standard error estimated by  RMLE (MWCMLE).
}
\details{
See \insertCite{weng2018improved;textual}{equivalenceTest}.
}
\references{
{
\insertRef{weng2018improved}{equivalenceTest}
}
}
