% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{histWNormDensity}
\alias{histWNormDensity}
\title{Histogram with a fitted normal density function}
\usage{
histWNormDensity(x, main = "")
}
\arguments{
\item{x}{the data}

\item{main}{the title of the plot}
}
\description{
Provide a histogram with a fitted normal density.
}
\examples{
x = rnorm(20)
histWNormDensity(x)

}
