% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{summarizeSample}
\alias{summarizeSample}
\title{Summarize data for equivalence test}
\usage{
summarizeSample(vecT, labelT, vecR, labelR, vecR1 = NULL, labelR1 = "")
}
\arguments{
\item{vecT}{vector of sample for T}

\item{labelT}{label for T}

\item{vecR}{vector of sample for R}

\item{labelR}{label for R}

\item{vecR1}{vector of sample for R1}

\item{labelR1}{label for R1}
}
\value{
a data.frame consisting the sample size, min, max, mean, SD, and percentage coefficient of variation for the samples
}
\description{
Summarize data for equivalence test, can be two datasets or three datasets.
}
\examples{
vecT = rnorm(10,-1.5,1)
vecR = rnorm(10)
vecR1 = rnorm(15,1,2)
ss = summarizeSample(vecT,"T",vecR,"R",vecR1,"R1")
}
